+++ .eslintrc.json
@@ -1,30 +1,25 @@
 {
-    "env": {
-        "browser": true,
-        "es2021": true
+  "env": {
+    "browser": true,
+    "es2021": true
+  },
+  "extends": [
+    "plugin:react/recommended",
+    "prettier",
+    "plugin:react/jsx-runtime"
+  ],
+  "parserOptions": {
+    "ecmaFeatures": {
+      "jsx": true
     },
-    "extends": [
-        "plugin:react/recommended",
-        "airbnb",
-        "prettier",
-        "plugin:react/jsx-runtime"
-    ],
-    "parserOptions": {
-        "ecmaFeatures": {
-            "jsx": true
-          },
-        "ecmaVersion": "latest",
-        "sourceType": "module"
-    },
-    "plugins": ["react"],
-    "rules": {
-        "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
-        "react/prop-types": 0,
-        "react/destructuring-assignment": [0],
-        "import/prefer-default-export": [0],
-        "default-param-last": [0],
-        "react/function-component-definition": [0],
-        "arrow-body-style": 0
-
-    }
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "plugins": ["react"],
+  "rules": {
+    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
+    "react/prop-types": 0,
+    "react/destructuring-assignment": [0],
+    "react/react-in-jsx-scope": "off"
+  }
 }
