+++ package.json
@@ -1,7 +1,7 @@
 {
-  "name": "sky-marketplace",
+  "name": "3a79971f-9aa",
   "version": "0.1.0",
-  "private": false,
+  "private": true,
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.6",
     "@testing-library/jest-dom": "^5.17.0",
@@ -9,19 +9,21 @@
     "@testing-library/user-event": "^13.5.0",
     "classnames": "^2.3.2",
     "react": "^18.2.0",
+    "react-audio-player": "^0.17.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.1.2",
     "react-router-dom": "^6.20.1",
     "react-scripts": "5.0.1",
-    "styled-components": "^6.1.1",
+    "styled-components": "^6.0.7",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
+    "eject": "react-scripts eject",
     "format": "prettier —write .",
-    "eject": "react-scripts eject"
+    "lint:js": "eslint ./src/**/*.js"
   },
   "eslintConfig": {
     "extends": [
@@ -42,14 +44,9 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^8.55.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-import": "^2.29.0",
-    "eslint-plugin-jsx-a11y": "^6.8.0",
-    "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "3.1.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-prettier": "^5.0.0",
+    "prettier": "^3.0.1",
     "stylelint": "^15.10.2",
     "stylelint-config-standard": "^34.0.0",
     "stylelint-prettier": "^4.0.2"
