+++ src/Components/Footer/Footer.styled.js
@@ -1,51 +1,40 @@
-import styled from "styled-components";
+import styled from 'styled-components';
 
 export const FooterWrapper = styled.footer`
-    display: none;
+  display: none;
 
-    @media screen and (max-width: 590px) {
+  @media screen and (max-width: 620px) {
     position: fixed;
     bottom: 0;
     left: 0;
     z-index: 3;
-    display: -webkit-box;
-    display: -ms-flexbox;
     display: flex;
-    -webkit-box-align: center;
-        -ms-flex-align: center;
-            align-items: center;
-    -webkit-box-pack: center;
-        -ms-flex-pack: center;
-            justify-content: center;
+    align-items: center;
+    justify-content: center;
     height: 54px;
     width: 100%;
-    background-color: #FFFFFF;
-    -webkit-box-shadow: 0px 4px 25px rgba(0, 0, 0, 0.1);
+    background-color: #ffffff;
     box-shadow: 0px 4px 25px rgba(0, 0, 0, 0.1);
-    }
-`
+  }
+`;
 
 export const FooterContainer = styled.div`
-@media screen and (max-width: 590px) {
+  @media screen and (max-width: 620px) {
     width: 225px;
     display: -webkit-box;
     display: -ms-flexbox;
     display: flex;
-    -webkit-box-align: center;
-        -ms-flex-align: center;
-            align-items: center;
-    -webkit-box-pack: justify;
-        -ms-flex-pack: justify;
-            justify-content: space-between;
-}
-`
+    align-items: center;
+    justify-content: space-between;
+  }
+`;
 
 export const FooterImg = styled.img`
-@media screen and (max-width: 590px) {
+  @media screen and (max-width: 620px) {
     width: 42px;
     height: 42px;
-}
-`
+  }
+`;
 // была ссылка внутри img
 // .footer__img img {
 //     width: 100%;
