+++ src/Pages/SellerProfile/SellerProfile.jsx
@@ -1,37 +1,18 @@
+import { HeaderSecond } from '../../Components/HeaderSecond/HeaderSecond';
+import { ReturnToMain } from '../../Components/ReturnToMain.js/ReturnToMain';
 import { Card } from '../../Components/Card/Card';
 import { Footer } from '../../Components/Footer/Footer';
-import { Link } from 'react-router-dom';
 import * as S from './SellerProfile.styled';
 
 export const SellerProfile = () => {
   return (
     <S.Wrapper>
       <S.Container>
-        <S.Header>
-          <S.HeaderNav>
-            <S.HeaderLogo>
-              <S.LogoMobLink target='_blank'>
-                <S.LogoMobImg src='img/logo-mob.png' alt='logo' />
-              </S.LogoMobLink>
-            </S.HeaderLogo>
-            <S.HeaderBtnPutAd>Разместить объявление</S.HeaderBtnPutAd>
-            <S.HeaderButtonSecond>Личный кабинет</S.HeaderButtonSecond>
-          </S.HeaderNav>
-        </S.Header>
-
+        <HeaderSecond />
         <S.Main>
           <S.MainContainer>
             <S.MainCenterBlock>
-              <S.Menu>
-                <S.MenuLogoLink>
-                  <S.MenuLogoImg src='img/logo.png' alt='logo' />
-                </S.MenuLogoLink>
-                <S.MenuForm action='#'>
-                  <Link to='/'>
-                    <S.MenuBtn>Вернуться на&nbsp;главную</S.MenuBtn>
-                  </Link>
-                </S.MenuForm>
-              </S.Menu>
+              <ReturnToMain />
               <S.MainH2>Профиль продавца</S.MainH2>
               <S.MainProfileSell>
                 <S.ProfileSellContent>
@@ -56,7 +37,7 @@ export const SellerProfile = () => {
                         </S.SellerImgMob>
                       </S.SellerImgMobBlock>
 
-                      <S.SellerBtn className='seller__btn btn-hov02'>
+                      <S.SellerBtn>
                         Показать&nbsp;телефон
                         <S.SellerBtnSpan>
                           8&nbsp;905&nbsp;ХХХ&nbsp;ХХ&nbsp;ХХ
